consolidate files, final renames
