+++ src/app/extensions/extension.config.ts
@@ -23,11 +23,10 @@
  * along with Alfresco. If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { RuleRef } from './rules/rule-ref';
-import { ActionRef } from './action-ref';
-import { ContentActionExtension } from './content-action.extension';
 import { NavBarGroupRef } from './navbar.extensions';
 import { RouteRef } from './routing.extensions';
+import { RuleRef } from './rule.extensions';
+import { ActionRef, ContentActionRef } from './action.extensions';
 
 export interface ExtensionConfig {
     version: string;
@@ -37,13 +36,13 @@ export interface ExtensionConfig {
     actions?: Array<ActionRef>;
     features?: {
         [key: string]: any;
-        create?: Array<ContentActionExtension>;
+        create?: Array<ContentActionRef>;
         viewer?: {
-            openWith?: Array<ContentActionExtension>;
+            openWith?: Array<ContentActionRef>;
         };
         navbar?: Array<NavBarGroupRef>;
         content?: {
-            actions?: Array<ContentActionExtension>;
+            actions?: Array<ContentActionRef>;
         };
     };
 }
